+++ CHANGELOG.md
@@ -191,6 +191,7 @@ here is a non-exhaustive list of significant departures from the original gem:
 * {OpenHAB::DSL::Rules::Terse terse rules} now have an `on_load` parameter.
 * {Item#all_groups Item#all_groups}, {Enumerable#all_members}, {Enumerable#groups}, {Enumerable#all_groups}.
 * {GenericItem#formatted_state GenericItem#formatted_state}.
+* {OpenHAB::DSL.transform} now available at top-level, like Rules DSL.
 
 ### Bug Fixes